+++ site/src/components/PageHeader.tsx
@@ -83,7 +83,7 @@ const PageHeader: React.SFC<PageHeaderProps> = ({
                 <Heading
                   variant={headingOnly ? "display-0" : "heading-1"}
                   tag="h1"
-                  classNameAndIHaveSpokenToDST={styles.headingTextOverrides}
+                  classNameOverride={styles.headingTextOverrides}
                 >
                   {headingText}
                 </Heading>